requirements1.txt
# Core Framework
streamlit==1.39.0

# Image Processing
opencv-python-headless==4.10.0.84
Pillow==10.4.0
scikit-image==0.24.0
numpy==1.26.4

# Deep Learning (for AI-based KYC)
tensorflow==2.17.0
keras==3.4.1
deepface==0.0.93

# Data Handling
pandas==2.2.3
openpyxl==3.1.5

# PDF Report Generation
reportlab==4.2.2

# Optional (recommended for smooth deployment)
protobuf==5.27.2
typing_extensions==4.12.2